+++ build_runner/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Remove `OnDeleteWriter`, add functionality to `ReaderWriter`.
 - Add `NodeType` to `AssetNode`, remove subtypes. Make mutations explicit.
 - Use `built_value` for `AssetNode` and related types.
+- Add details of what changed and what is built to `--verbose` logging.
 
 ## 2.4.15